+++ apps/common/inputparser.h
@@ -16,7 +16,7 @@ class InputParser
 public:
   InputParser(int& argc, char** argv);
 
-  int getCmdOption(const std::string& option, std::vector< std::string >& value) const;
+  size_t getCmdOption(const std::string& option, std::vector< std::string >& value) const;
   bool cmdOptionExists(const std::string& option) const;
 
 private:
